// ts/enum.ts
export enum LOGIN_TYPE {
  PATIENT = "Patient",
  CLINICIAN = "Clinician",
}

export const unitEnum = {
  STRING: "string",
  MILI_GRAM: "mg",
  POUND: "lbs",
  FEET: "ft",
  INCH: "in",
  NUMBER: "NUMBER",
};

export enum layoutModeEnum {
  TEXT_ANSWER = "TEXT_ANSWER",
  AUDIO = "AUDIO",
  AVATAR = "AVATAR",
  CONDITIONAL = "CONDITIONAL",
  AVATAR_MODE = "avatar-mode",
  TEXT_MODE = "text-mode",
}